Nitpicks

+++ data_structures_and_algorithms/10_mathematical/08_02_extract_power_of_number_for_base_with_division_of_logarithms.py
@@ -19,6 +19,6 @@ if __name__ == '__main__':
     numerator_log = math.log(number)
     denominator_log = math.log(base)
 
-    exponent = int(numerator_log / denominator_log)
+    exponent = numerator_log / denominator_log
 
     print(f"{base} ** {exponent} = {number}")
